+++ components/metrics/ActiveUsers.js
@@ -5,7 +5,7 @@ import useFetch from 'hooks/useFetch';
 import Dot from 'components/common/Dot';
 import styles from './ActiveUsers.module.css';
 
-export default function ActiveUsers({ websiteId, className, value, interval = 60000 }) {
+export default function ActiveUsers({ websiteId, className, value, interval = 60000000 }) {
   const url = websiteId ? `/website/${websiteId}/active` : null;
   const { data } = useFetch(url, {
     interval,
