+++ lib/constants.js
@@ -66,6 +66,7 @@ export const EVENT_COLORS = [
 
 export const POSTGRESQL = 'postgresql';
 export const MYSQL = 'mysql';
+export const CLICKHOUSE = 'clickhouse';
 
 export const MYSQL_DATE_FORMATS = {
   minute: '%Y-%m-%d %H:%i:00',
@@ -81,6 +82,14 @@ export const POSTGRESQL_DATE_FORMATS = {
   day: 'YYYY-MM-DD',
   month: 'YYYY-MM-01',
   year: 'YYYY-01-01',
+};
+
+export const CLICKHOUSE_DATE_FORMATS = {
+  minute: '%Y-%m-%d %H:%M:00',
+  hour: '%Y-%m-%d %H:00:00',
+  day: '%Y-%m-%d',
+  month: '%Y-%m-01',
+  year: '%Y-01-01',
 };
 
 export const DOMAIN_REGEX =
