+++ package.json
@@ -60,6 +60,7 @@
     "chalk": "^4.1.1",
     "chart.js": "^2.9.4",
     "classnames": "^2.3.1",
+    "clickhouse": "^2.5.0",
     "colord": "^2.9.2",
     "cors": "^2.8.5",
     "cross-spawn": "^7.0.3",
