+++ queries/admin/account/updateAccount.js
@@ -1,5 +1,4 @@
-import { runQuery } from 'lib/queries';
-import prisma from 'lib/db';
+import { prisma, runQuery } from 'lib/db';
 
 export async function updateAccount(user_id, data) {
   return runQuery(
