+++ queries/admin/website/getWebsiteByShareId.js
@@ -1,5 +1,4 @@
-import { runQuery } from 'lib/queries';
-import prisma from 'lib/db';
+import { prisma, runQuery } from 'lib/db';
 
 export async function getWebsiteByShareId(share_id) {
   return runQuery(
