+++ queries/analytics/event/saveEvent.js
@@ -1,5 +1,5 @@
-import { runQuery } from 'lib/queries';
-import prisma from 'lib/db';
+import { prisma, runQuery } from 'lib/db';
+
 import { URL_LENGTH } from 'lib/constants';
 
 export async function saveEvent(website_id, session_id, url, event_type, event_value) {
