+++ queries/analytics/pageview/getPageviewMetrics.js
@@ -1,4 +1,4 @@
-import { parseFilters, rawQuery } from 'lib/queries';
+import { parseFilters, rawQuery } from 'lib/db';
 
 export function getPageviewMetrics(website_id, start_at, end_at, field, table, filters = {}) {
   const params = [website_id, start_at, end_at];
